Rename Introduction-to-Git to Git
If I'm going to have a whole directory, might as well structure it
accordingly. I think it would be a good idea to have lessons by major
topic. In the Git folder, I'll eventually change the README to a table
of contents, which will be what people see when they browse there. The
remaining files will be different chapters/posts in the tutorial -- for
instance, my introduction will be the first post; I think I'll do a
separate file for GitHub. And a few others for more advanced Git
features, etc.